fix(pos): resolve ESLint warnings in POS components
  Extract useShift hook to its own file to fix fast-refresh violation in
  ShiftGuard, and fix DiscountDialog useEffect by inlining the null check
  (removing non-null assertion) and completing the dependency array.
