+++ src/views/pos/components/useShift.ts
@@ -1,0 +1,12 @@
+import { createContext, useContext } from 'react'
+import type { Shift } from '@/services/POSService'
+
+export const ShiftContext = createContext<Shift | null>(null)
+
+export const useShift = () => {
+    const shift = useContext(ShiftContext)
+    if (!shift) {
+        throw new Error('useShift must be used within ShiftGuard')
+    }
+    return shift
+}
